+++ src/main/kotlin/org/radarbase/push/integration/garmin/util/offset/OffsetPersistenceFactory.kt
@@ -1,4 +1,4 @@
-package org.radarbase.push.integration.common.redis.offset
+package org.radarbase.push.integration.garmin.util.offset
 
 import java.nio.file.Path
