+++ src/main/kotlin/org/radarbase/push/integration/garmin/util/offset/Offsets.kt
@@ -1,4 +1,4 @@
-package org.radarbase.push.integration.common.redis.offset
+package org.radarbase.push.integration.garmin.util.offset
 
 import java.time.Instant
 import java.util.concurrent.ConcurrentHashMap
